Moved language state and tests
