+++ app/containers/LanguageProvider/index.js
@@ -12,7 +12,7 @@ import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
 import { IntlProvider } from 'react-intl';
 
-import { makeSelectLocale } from './selectors';
+import { makeSelectLocale } from '../../state/language/selectors';
 
 export class LanguageProvider extends React.PureComponent {
   // eslint-disable-line react/prefer-stateless-function
