+++ app/tests/state/route/selectors.test.js
@@ -1,4 +1,4 @@
-import { makeSelectLocation } from 'containers/App/selectors';
+import { makeSelectLocation } from '../../../state/route/selectors';
 
 describe('makeSelectLocation', () => {
   it('should select the location', () => {
